Add mouse wheel handling trough InputFrame(s)
Add what is needed in order to let InputFrame handle
mouse wheel, which is then used by ShipViewController in its InputFrame.

Prepare SectorView to be moved on a InputFrame pattern.

Skip (for now) mouse wheel handling in LuaInput avoiding crashes,
this could be improved changing mouse wheel to be an axis, which
should be easier on future changes.