Move Clamp and DE2RAD function in 'utils.h'
"libs.h" is a bin with a collection of headers, while "utils.h"
is a place where code is. Thus it seems natural to place these
functions in that place.

In addition to this, some headers which uses "libs.h" is decoupled,
improving parsing time and giving a better idea of code complexity.

One more thing is that some include in Scenegraph went decouple using
forward declarations and thus reducing incremental builds time.

Last ShieldRenderParameters is in its header, which is included
in render and in shield code, avoiding code duplications