Check collision against volumes (CSG geometry)
Define a couple of primitive (CSG_*) which are used
by geoms in order to replace triangle collision tests.
Those CSG_*, when present in json definition, are
loaded and brought to geom.
If a flag is set, when an object is inside those volumes
it trigger docking.

In order to ease debug, that data is duplicated and sank
in model to be shown with collision meshes

Sides:
* A new drawable Box3D which is indexed and with only 8 vertex
  (no normals or UV but quite fast)
* Remove scenegraph/Scenegraph.h, a bin included a lot and
  decouple files
* Decouple some files forwarding Model.h instead of including
* Use more unique_ptr instead of naked pointers